+++ changelog.md
@@ -1,5 +1,7 @@
 **WARNING: YOUR CORE CONFIG FILE WILL BE OVERWRITTEN**
 ### Changelog
 
-- use xrayhelper v1.2.5
+- use xrayhelper v1.3.0
+- rename clash.meta
+- remove clash and clash.premium
 - update README